+++ src/Twilio/Http/CurlClient.php
@@ -42,10 +42,14 @@ class CurlClient implements Client {
             }
 
             $parts = \explode("\r\n\r\n", $response, 3);
-            list($head, $body) = (\preg_match('/\AHTTP\/1.\d 100 Continue\Z/', $parts[0])
-                               || \preg_match('/\AHTTP\/1.\d 200 Connection established\Z/', $parts[0]))
-                               ? array($parts[1], $parts[2])
-                               : array($parts[0], $parts[1]);
+
+            list($head, $body) = (
+                \preg_match('/\AHTTP\/1.\d 100 Continue\Z/', $parts[0])
+                || \preg_match('/\AHTTP\/1.\d 200 Connection established\Z/', $parts[0])
+                || \preg_match('/\AHTTP\/1.\d 200 Tunnel established\Z/', $parts[0])
+            )
+                ? array($parts[1], $parts[2])
+                : array($parts[0], $parts[1]);
 
             if ($this->debugHttp) {
                 $u = \parse_url($url);