+++ src/Twilio/Rest/Serverless/V1/Service/Environment/LogOptions.php
@@ -19,10 +19,14 @@ abstract class LogOptions {
     /**
      * @param string $functionSid The SID of the function whose invocation produced
      *                            the Log resources to read
+     * @param \DateTime $startDate The date and time after which the Log resources
+     *                             must have been created.
+     * @param \DateTime $endDate The date and time before which the Log resource
+     *                           must have been created.
      * @return ReadLogOptions Options builder
      */
-    public static function read($functionSid = Values::NONE) {
-        return new ReadLogOptions($functionSid);
+    public static function read($functionSid = Values::NONE, $startDate = Values::NONE, $endDate = Values::NONE) {
+        return new ReadLogOptions($functionSid, $startDate, $endDate);
     }
 }
 
@@ -30,9 +34,15 @@ class ReadLogOptions extends Options {
     /**
      * @param string $functionSid The SID of the function whose invocation produced
      *                            the Log resources to read
+     * @param \DateTime $startDate The date and time after which the Log resources
+     *                             must have been created.
+     * @param \DateTime $endDate The date and time before which the Log resource
+     *                           must have been created.
      */
-    public function __construct($functionSid = Values::NONE) {
+    public function __construct($functionSid = Values::NONE, $startDate = Values::NONE, $endDate = Values::NONE) {
         $this->options['functionSid'] = $functionSid;
+        $this->options['startDate'] = $startDate;
+        $this->options['endDate'] = $endDate;
     }
 
     /**
@@ -44,6 +54,30 @@ class ReadLogOptions extends Options {
      */
     public function setFunctionSid($functionSid) {
         $this->options['functionSid'] = $functionSid;
+        return $this;
+    }
+
+    /**
+     * The date/time (in GMT, ISO 8601) after which the Log resources must have been created. Defaults to 1 day prior to current date/time.
+     *
+     * @param \DateTime $startDate The date and time after which the Log resources
+     *                             must have been created.
+     * @return $this Fluent Builder
+     */
+    public function setStartDate($startDate) {
+        $this->options['startDate'] = $startDate;
+        return $this;
+    }
+
+    /**
+     * The date/time (in GMT, ISO 8601) before which the Log resources must have been created. Defaults to current date/time.
+     *
+     * @param \DateTime $endDate The date and time before which the Log resource
+     *                           must have been created.
+     * @return $this Fluent Builder
+     */
+    public function setEndDate($endDate) {
+        $this->options['endDate'] = $endDate;
         return $this;
     }