+++ tests/Twilio/Integration/Autopilot/V1/Assistant/QueryTest.php
@@ -57,7 +57,8 @@ class QueryTest extends HolodeckTestCase {
                 "url": "https://autopilot.twilio.com/v1/Assistants/UAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Queries/UHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "account_sid": "ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "sid": "UHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-                "source_channel": "voice"
+                "source_channel": "voice",
+                "dialogue_sid": "UKaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
             }
             '
         ));
@@ -136,7 +137,8 @@ class QueryTest extends HolodeckTestCase {
                         "url": "https://autopilot.twilio.com/v1/Assistants/UAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Queries/UHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                         "account_sid": "ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                         "sid": "UHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-                        "source_channel": null
+                        "source_channel": null,
+                        "dialogue_sid": "UKaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
                     }
                 ],
                 "meta": {
@@ -203,7 +205,8 @@ class QueryTest extends HolodeckTestCase {
                 "url": "https://autopilot.twilio.com/v1/Assistants/UAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Queries/UHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "account_sid": "ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "sid": "UHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-                "source_channel": "voice"
+                "source_channel": "voice",
+                "dialogue_sid": null
             }
             '
         ));
@@ -255,7 +258,8 @@ class QueryTest extends HolodeckTestCase {
                 "url": "https://autopilot.twilio.com/v1/Assistants/UAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Queries/UHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "account_sid": "ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "sid": "UHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-                "source_channel": "sms"
+                "source_channel": "sms",
+                "dialogue_sid": "UKaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
             }
             '
         ));