+++ tests/Twilio/Unit/Http/GuzzleClientTest.php
@@ -40,7 +40,7 @@ final class GuzzleClientTest extends UnitTest
     {
         $this->mockHandler->append(new Response());
         $response = $this->client->request('post', 'https://www.whatever.com', ['myquerykey' => 'myqueryvalue'], ['myparamkey' => 'myparamvalue']);
-        $this->assertSame(null, $response->getContent());
+        $this->assertNull($response->getContent());
         $this->assertSame(200, $response->getStatusCode());
         $this->assertSame([], $response->getHeaders());
 
@@ -55,7 +55,7 @@ final class GuzzleClientTest extends UnitTest
     {
         $this->mockHandler->append(new BadResponseException('Not found', new Request('get', 'https://www.whatever.com'), new Response(404)));
         $response = $this->client->request('post', 'https://www.whatever.com', ['myquerykey' => 'myqueryvalue'], ['myparamkey' => 'myparamvalue']);
-        $this->assertSame(null, $response->getContent());
+        $this->assertNull($response->getContent());
         $this->assertSame(404, $response->getStatusCode());
         $this->assertSame([], $response->getHeaders());
 
@@ -71,5 +71,14 @@ final class GuzzleClientTest extends UnitTest
         $this->mockHandler->append(new RequestException('Not found', new Request('get', 'https://www.whatever.com')));
         $this->expectException(HttpException::class, 'Unable to complete the HTTP request');
         $this->client->request('post', 'https://www.whatever.com', ['myquerykey' => 'myqueryvalue'], ['myparamkey' => 'myparamvalue']);
+    }
+
+    public function testQueryParams()
+    {
+        $this->mockHandler->append(new Response());
+        $this->client->request('get', 'https://www.whatever.com?foo=bar');
+        $request = $this->mockHandler->getLastRequest();
+
+        $this->assertSame('https://www.whatever.com?foo=bar', (string)$request->getUri());
     }
 }