+++ tests/Twilio/Unit/Rest/ClientTest.php
@@ -3,6 +3,8 @@
 
 namespace Twilio\Tests\Unit\Rest;
 
+use Twilio\Exceptions\ConfigurationException;
+use Twilio\Exceptions\TwilioException;
 use Twilio\Http\CurlClient;
 use Twilio\Http\Response;
 use Twilio\Rest\Client;
@@ -12,25 +14,19 @@ use Twilio\Tests\Unit\UnitTest;
 
 class ClientTest extends UnitTest {
 
-    /**
-     * @expectedException \Twilio\Exceptions\ConfigurationException
-     */
-    public function testThrowsWhenUsernameAndPasswordMissing() {
-        new Client(null, null, null, null, null, array());
+    public function testThrowsWhenUsernameAndPasswordMissing(): void {
+        $this->expectException(ConfigurationException::class);
+        new Client(null, null, null, null, null, []);
     }
 
-    /**
-     * @expectedException \Twilio\Exceptions\ConfigurationException
-     */
-    public function testThrowsWhenUsernameMissing() {
-        new Client(null, 'password', null, null, null, array());
+    public function testThrowsWhenUsernameMissing(): void {
+        $this->expectException(ConfigurationException::class);
+        new Client(null, 'password', null, null, null, []);
     }
 
-    /**
-     * @expectedException \Twilio\Exceptions\ConfigurationException
-     */
-    public function testThrowsWhenPasswordMissing() {
-        new Client('username', null, null, null, null, array());
+    public function testThrowsWhenPasswordMissing(): void {
+        $this->expectException(ConfigurationException::class);
+        new Client('username', null, null, null, null, []);
     }
 
     public function testUsernamePulledFromEnvironment() {
@@ -123,10 +119,8 @@ class ClientTest extends UnitTest {
 		$client->validateSslCertificate($curlClient);
 	}
 
-	/**
-	 * @expectedException \Twilio\Exceptions\TwilioException
-	 */
 	public function testValidationSslCertificateError() {
+        $this->expectException(TwilioException::class);
 		$client = new Client('username', 'password');
 		$curlClient = $this->createMock(CurlClient::class);
 		$curlClient
@@ -134,7 +128,7 @@ class ClientTest extends UnitTest {
             ->method('request')
 			->willReturn(new Response(504, ''))
         ;
-		
+
 		$client->validateSslCertificate($curlClient);
 	}